+++ wdfwd/tests/test_tail.py
@@ -59,7 +59,7 @@ def _ktail():
     bdir = finfo['dir']
     ptrn = finfo['pattern']
     tag = finfo['tag']
-    kcfg = KinesisCfg(KN_TEST_STREAM, None, None, None)
+    kcfg = KinesisCfg(KN_TEST_STREAM, 'ap-northeast-1', None, None)
     tail = FileTailer(bdir, ptrn, tag, pos_dir, kcfg,
                       send_term=0, update_term=0, echo=True,
                       max_between_data=100 * 100)
@@ -144,6 +144,9 @@ def test_tail_file_kinesis(rmlogs, ktail):
     # send previous data if it has moderate size
     assert ktail.get_sent_pos() == 0
     assert ktail.may_send_newlines() == 1
+    knc = ktail.kclient
+    #for rec in iter_kinesis_records(knc, ktail.ksent_shid, ktail.ksent_seqn):
+        #assert 'ts_' in rec
 
     with open(path, 'a') as f:
         f.write('2\n')
@@ -167,16 +170,12 @@ def test_tail_file_kinesis(rmlogs, ktail):
 
     # stress test
     with open(path, 'a') as f:
-        for i in xrange(4, 1000):
+        for i in xrange(4, 100):
             f.write('{}\n'.format(i))
-    assert ktail.may_send_newlines() == 996
+    assert ktail.may_send_newlines() == 96
 
-
-    knc = ktail.kclient
-    shid = ktail.ksent_shid
-    seqn = ktail.ksent_seqn
-    for rec in iter_kinesis_records(knc, shid, seqn):
-        assert 'ts_' in rec
+    #for rec in iter_kinesis_records(knc, ktail.ksent_shid, ktail.ksent_seqn):
+        #assert 'ts_' in rec
 
 
 def test_tail_file_rotate(rmlogs, ftail):